+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [v1.1.139] - 2024-06-21
+
+### Added
+
+- Support of `ext_chromium_mv3` platform
+
+### Fixed
+
+- Do not build filters due to `platformsIncluded` and `platformsExcluded`
+  if limited list of filters are specified to build
+
+[v1.1.139]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.1.138...v1.1.139
+
+
 ## [v1.1.138] - 2024-06-21
 
 ### Added
@@ -13,18 +28,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Return the rules in their original order after applying the `@include` directive option `/optimiseDomainBlockingRules` [#217]
+- Return the rules in their original order after applying the `@include` directive option `/optimizeDomainBlockingRules` [#217]
 
 [v1.1.138]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.1.136...v1.1.138
 
-## [v1.1.136] - 2024-05-22
 
+## [v1.1.136] - 2024-05-22
 
 ### Changed
 
 - Updated [@adguard/filters-downloader] to v2.2.1
 
-[v1.1.135]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.1.135...v1.1.136
+[v1.1.136]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.1.135...v1.1.136
 
 
 ## [v1.1.135] - 2024-05-20